feat: 메뉴 변경 패칭 함수 공통 hook으로 분리
