+++ app/hooks/useClickActive.tsx
@@ -1,7 +1,17 @@
+import { useCallback } from "react";
+import { useLocation, useNavigate, useOutletContext } from "@remix-run/react";
+
 import { useMap } from "~/shared/contexts/Map";
+import { IRegister, IReview } from "~/shared/types";
+import { useFetch } from "./useFetch";
 
 export function useClickActive() {
-  const { GNB, setGNB } = useMap();
+  const navigate = useNavigate();
+  const location = useLocation();
+  const { user } = useOutletContext<{ user: IRegister }>();
+
+  const { GNB, setGNB, mapData } = useMap();
+  const { fetchCafeData } = useFetch();
 
   const handleActive = (id: string) => {
     setGNB(
@@ -11,5 +21,17 @@ export function useClickActive() {
     );
   };
 
-  return { handleActive };
+  const handleMenu = useCallback(
+    (id: string, userReview: IReview[] | null) => {
+      if (id !== "default" && user === null) {
+        navigate("/signin");
+      } else {
+        handleActive(id);
+        fetchCafeData(id, userReview);
+      }
+    },
+    [navigate, user, mapData, location]
+  );
+
+  return { handleActive, handleMenu };
 }
